Product test coverage. Fix broken validation of metadata section (deprecate explicit product name propagation)
